PUBDEV-7037: Added rgs filtering support for CartesianWalker. Moved some tests to h2o-algos module as TE is extension module.
